+++ src/SumatraPDF.h
@@ -236,6 +236,7 @@ constexpr u32 kUiSidebarDirty = 0x20; // repaint toc/favorites boxes and their s
 
 void ScheduleUiUpdate(MainWindow* win, u32 flags = kUiRelayout, int sidebarDx = -1);
 void DuplicateTabInNewWindow(WindowTab* tab);
+void DuplicateTabInNewTab(WindowTab* tab);
 void CopyFilePath(WindowTab*);
 
 MainWindow* FindMainWindowByFile(Str file, bool focusTab, MainWindow* limitWin = nullptr);